projects/de-event-analytics-pipeline/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.3
